refactor: use the error handlers from common.sh in load.sh
load.sh had its own copies of errorAndExit and terminateAndExit. clean
now skips the temporary directory when the calling script sets no
TEMP_DIR, as load.sh does.

+++ pipelines/lib/common.sh
@@ -134,8 +134,11 @@ log() { echo "$(date +'[%s (%F %T)]')" "$@"; }
 #   None                                                                       #
 ################################################################################
 clean() {
-	# Clean contents of temporary directory
-	rm -rf "${TEMP_DIR:?}/$UNIPEPT_TEMP_CONSTANT"
+	# Clean contents of temporary directory, if the calling script uses one
+	if [[ -n "${TEMP_DIR:-}" ]]
+	then
+		rm -rf "${TEMP_DIR:?}/$UNIPEPT_TEMP_CONSTANT"
+	fi
 	export TMPDIR="$OLD_TMPDIR"
 }
